+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [4.5.2](https://github.com/xdy/twodsix-foundryvtt/compare/v4.5.1...v4.5.2) (2023-11-15)
+
+
+### Bug Fixes
+
+* AE not active for linked actors on load ([#1498](https://github.com/xdy/twodsix-foundryvtt/issues/1498)) ([8026f2e](https://github.com/xdy/twodsix-foundryvtt/commit/8026f2ec9c181bee43b0708aabe6af25fdcb420a))
+
+## [4.5.1](https://github.com/xdy/twodsix-foundryvtt/compare/v4.5.0...v4.5.1) (2023-11-14)
+
+
+### Bug Fixes
+
+* sort skills names alphabetically in select lists  ([#1497](https://github.com/xdy/twodsix-foundryvtt/issues/1497)) ([44b1f7f](https://github.com/xdy/twodsix-foundryvtt/commit/44b1f7fb54ec4ac16201d4b366b5e4e6dd37d48e))
+
 # [4.5.0](https://github.com/xdy/twodsix-foundryvtt/compare/v4.4.0...v4.5.0) (2023-11-11)